Expatriate the inline string methods
