+++ src/d/d_rawarchive.cpp
@@ -1,8 +1,6 @@
 #include <d/d_rawarchive.h>
+#include <inline_string.h>
 #include <rvl/VI.h>
-// clang-format off
-#include <MSL_C/string.h>
-// clang-format on
 
 class UnkManager {
 public:
@@ -131,32 +129,6 @@ bool dRawArcEntry_c::loadArcFromDiskChecked(const char *fileName, const char *di
         return loadArcFromDisk(fileName, arcPath, mountDirection, heap);
     }
     return false;
-}
-
-// Regswap
-inline void inline_strncat(char *dest, const char *src, size_t destSize) {
-    if (src != nullptr) {
-        size_t destLen = strlen(dest);
-        size_t copyLen = strlen(src);
-
-        // Make sure copy length isnt more than destination length
-        if (destLen + copyLen + 1 >= destSize) {
-            copyLen = destSize - destLen - 1;
-        }
-
-        strncpy(dest + destLen, src, copyLen);
-
-        // make sure string is null terminated
-        size_t offset = destLen + copyLen;
-        dest[offset] = '\0';
-    }
-}
-
-inline void inline_strncpy(char *dest, const char *src, size_t destSize) {
-    if (src != dest) {
-        dest[0] = '\0';
-        inline_strncat(dest, src, destSize);
-    }
 }
 
 bool dRawArcEntry_c::loadArcFromDisk(const char *arcName, const char *arcPath, u8 mountDirection, EGG::Heap *heap) {
